+++ purchases_ui_flutter/android/src/main/kotlin/com/revenuecat/purchases_ui_flutter/views/PaywallView.kt
@@ -72,6 +72,14 @@ internal class PaywallView(
             override fun onRestoreError(error: Map<String, Any?>) {
                 methodChannel.invokeMethod("onRestoreError", error)
             }
+
+            override fun onWebCheckoutOpened() {
+                methodChannel.invokeMethod("onWebCheckoutOpened", null)
+            }
+
+            override fun onUrlOpened(url: String) {
+                methodChannel.invokeMethod("onUrlOpened", mapOf("url" to url))
+            }
         })
         // Custom variables must be set before setting the offering to ensure they're applied
         val customVariables = creationParams["customVariables"] as? Map<String, Any?>