+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 10.7.0
+## RevenueCat SDK
+### ✨ New Features
+* Enables support for multipage paywalls
+### 📦 Dependency Updates
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 18.28.0 (#1844) via RevenueCat Git Bot (@RCGitBot)
+  * [Android 10.16.0](https://github.com/RevenueCat/purchases-android/releases/tag/10.16.0)
+  * [iOS 5.83.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.83.0)
+  * [iOS 5.82.0](https://github.com/RevenueCat/purchases-ios/releases/tag/5.82.0)
+  * [Web 1.49.1](https://github.com/RevenueCat/purchases-js/releases/tag/1.49.1)
+* [AUTOMATIC BUMP] Updates purchases-hybrid-common to 18.27.0 (#1841) via RevenueCat Git Bot (@RCGitBot)
+
+### 🔄 Other Changes
+* Remove workflows dangerous settings (#1843) via Facundo Menzella (@facumenzella)
+
 ## 10.6.0
 ## RevenueCat SDK
 ### ✨ New Features